Добавялет незначительные изменения
